add fusionsphere support

+++ chef/cookbooks/openstack-common/recipes/default.rb
@@ -84,33 +84,70 @@ when 'suse'
       end
     end
   end
-  if node['lsb']['description'][/^SUSE Linux Enterprise Server/]
-    release, patchlevel = node['platform_version'].split('.')
-    zypp_release = "SLE_#{release}_SP#{patchlevel}"
-  elsif node['lsb']['description'][/^openSUSE/]
-    zypp_release = 'openSUSE_' + node['lsb']['release']
+
+  if node['local_repo'].nil? or node['local_repo'].empty?
+    if node['lsb']['description'][/^SUSE Linux Enterprise Server/]
+      release, patchlevel = node['platform_version'].split('.')
+      zypp_release = "SLE_#{release}_SP#{patchlevel}"
+    elsif node['lsb']['description'][/^openSUSE/]
+      zypp_release = 'openSUSE_' + node['lsb']['release']
+    end
+    zypp = node['openstack']['zypp']
+    repo_uri = zypp['uri'].gsub(
+      '%release%', node['openstack']['release'].capitalize)
+    repo_uri.gsub! '%suse-release%', zypp_release
+    repo_alias = 'Cloud:OpenStack:' + node['openstack']['release'].capitalize
+
+    # TODO(iartarisi) this should be moved to its own cookbook
+    bash 'add repository key' do
+      cwd '/tmp'
+      code <<-EOH
+        gpg --keyserver pgp.mit.edu --recv-keys #{zypp['repo-key']}
+        gpg --armor --export #{zypp['repo-key']} > cloud.asc
+        rpm --import cloud.asc
+        rm -f cloud.asc
+      EOH
+
+      not_if { Mixlib::ShellOut.new('rpm -qa gpg-pubkey*').run_command.stdout.include? zypp['repo-key'].downcase }
+    end
+
+    execute 'add repository' do
+      command "zypper addrepo -k --check #{repo_uri} #{repo_alias}"
+      not_if { Mixlib::ShellOut.new('zypper repos --export -').run_command.stdout.include? repo_uri }
+    end
+
+    execute 'add repository vendor' do
+      command "echo '[#{repo_alias}]\nvendors = #{repo_alias}' > /etc/zypp/vendors.d/#{repo_alias}"
+      not_if { ::File.exists?("/etc/zypp/vendors.d/#{repo_alias}") }
+    end
   end
-  zypp = node['openstack']['zypp']
-  repo_uri = zypp['uri'].gsub(
-    '%release%', node['openstack']['release'].capitalize)
-  repo_uri.gsub! '%suse-release%', zypp_release
-  repo_alias = 'Cloud:OpenStack:' + node['openstack']['release'].capitalize
 
-  # TODO(iartarisi) this should be moved to its own cookbook
-  bash 'add repository key' do
-    cwd '/tmp'
-    code <<-EOH
-      gpg --keyserver pgp.mit.edu --recv-keys #{zypp['repo-key']}
-      gpg --armor --export #{zypp['repo-key']} > cloud.asc
-      rpm --import cloud.asc
-      rm -f cloud.asc
-    EOH
+  if node['lsb']['codename'] == 'UVP'
+    package 'FSComponentUtil' do
+      options platform_options['package_overrides']
+      action :upgrade
+    end
 
-    not_if { Mixlib::ShellOut.new('rpm -qa gpg-pubkey*').run_command.stdout.include? zypp['repo-key'].downcase }
-  end
+    template '/etc/FSSecurity/keylist.ini' do
+      source 'keylist.ini.erb'
+      owner "root"
+      group "root"
+      mode 00755
+    end
 
-  execute 'add repository' do
-    command "zypper addrepo --check #{repo_uri} #{repo_alias}"
-    not_if { Mixlib::ShellOut.new('zypper repos --export -').run_command.stdout.include? repo_uri }
+    group 'openstack'
+    user 'openstack' do
+      shell "/bin/bash"
+      gid "openstack"
+      system true
+      supports :manage_home => false
+    end
+
+    directory "/home/openstack" do
+      owner "openstack"
+      group "openstack"
+      mode 0755
+      action :create
+    end
   end
 end
